rv-site/requirements.txt
amqp==2.6.1
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-appconf==1.0.4
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-clear-cache==0.3
django-cms==3.9.0
django-filer==2.0.2
django-formtools==2.2
django-google-analytics-app==5.0.3
django-ipware==3.0.7
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-menus==1.1.2
django-mptt==0.12.0
django-nested-admin==3.3.3
django-nested-inline==0.4.2
django-phone-field==1.8.1
django-polymorphic==3.0.0
django-recaptcha==2.0.6
django-richtextfield==1.6
django-sekizai==2.0.0
django-tinymce==3.3.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-forms-maintained==202011091152
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
easy-thumbnails==2.7.1
et-xmlfile==1.1.0
hashids==1.3.1
html2text==2020.1.16
html5lib==1.1
idna==2.10
install==1.3.4
jsonfield==3.1.0
kombu==4.6.11
openpyxl==3.0.7
Pillow==8.2.0
psycopg2-binary==2.9.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-monkey-business==1.0.0
pytz==2021.1
pyuploadcare==2.7.0
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
South==1.0.2
sqlparse==0.4.1
structlog==21.1.0
tablib==3.0.0
Unidecode==1.1.2
urllib3==1.26.4
vine==1.3.0
webencodings==0.5.1
